refactor(config): 拆 parseClaudeModelProfile 到 config-parse,修 mock 跨文件污染
claude-models/claude-agent-process/codex-models 三个测试文件用 mock.module('./config') 隔离 config 单例,但 bun 的模块 mock 无法被 mock.restore() 撤销 —— 显式多文件同批运行(bun test a.test.ts b.test.ts)时泄漏到同作用域其他文件,config.test.ts 真实导入的 parseClaudeModelProfile 拿到 mock 后的残缺对象而抛 SyntaxError(标准全量 bun test 因每文件隔离不受影响)。

把纯函数 parseClaudeModelProfile + ClaudeModelConfig type 抽到独立模块 src/config-parse.ts(不依赖 config 单例),config.ts re-export 保持对外 API 不变;config.test.ts 改从 ./config-parse 导入即不受污染。另 claude-models.test.ts 去掉 mock.module,改真实 config + beforeEach/afterEach 覆盖/恢复,与 claude-agent-process.test.ts 风格一致。

验证:3 文件同跑 72 pass(原 70 pass+1 fail+1 error)、全量 bun test 449 pass / 0 fail、bun run build 通过。
